Adding proper attribution to AI
I will finish this tmw, it's late and I don't have the time, tldr: this is an insane thing for me to try to do alone, so I have used Gen AI to understand how to do things and to write code. I will add proper attribution. In this, webscrapers will not be submitted as part of my Final Project.